requirements.txt
colorama==0.4.6
requests==2.32.3
pillow==11.0.0
urllib3==2.2.3